refactor(RootState): rename
