+++ src/app/store/root-state.interface.ts
@@ -5,6 +5,7 @@ import { PushNotificationState } from '../push-notification/push-notification-st
 import { SettingsState } from '../fitness-schedule/store/settings/settings-state';
 import { RouterState } from '@angular/router';
 import { AuthenticationState } from '../authentication/store/authentication-state.interface';
+import { SyncState } from '../sync/sync-state.interface';
 /**
  * Created by Kim on 02.04.2017.
  */
@@ -16,4 +17,5 @@ export interface IAppState {
   settings?: SettingsState;
   authentication?: AuthenticationState;
   router?: RouterState;
+  sync?: SyncState;
 }
